move method from dynamoUtilities to DynamoCore as internal method
